Split arrangers into another module
